+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - Plugins
     - AutoLogin -- All responses now update the framework cookies.
     - Proxy -- Fixed out-of-scope error for `TemplateScope` helper class.
+- Modules
+    - Audit
+        - Path traversal (`path_traversal`)
+            - Added more fingerprints for `/etc/passwd`.
 
 ## Version 0.4.3 _(July 06, 2013)_